requirements.txt
# Prizolov Agent OS - Dependencies

# Для типизации
typing-extensions>=4.0.0

# Для работы с .env файлами
python-dotenv>=1.0.0

# Для тестирования
pytest>=7.0.0
pytest-cov>=4.0.0

# Для линтинга и форматирования
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0

# Для RAG (опционально)
# chromadb>=0.4.0
# sentence-transformers>=2.2.0
